+++ examples/with-next-rainbow-button/CHANGELOG.md
@@ -1,5 +1,11 @@
 # with-next-rainbow-button
 
+## 0.0.7
+
+### Patch Changes
+
+- @rainbow-me/rainbow-button@0.1.5
+
 ## 0.0.6
 
 ### Patch Changes